Add task7

+++ 0x00-MySQL_Advanced/6-bonus.sql
@@ -6,7 +6,7 @@ DELIMITER //
 CREATE PROCEDURE AddBonus(
 	IN user_id INT,
 	IN project_name VARCHAR(255),
-	IN score INT
+	IN score DECIMAL(10, 2)
 )
 BEGIN
 	DECLARE project_id INT;
